some minor cleanup, dialogs now can be centered on the pointer if the user prefers this

+++ util/string_view.h
@@ -393,7 +393,7 @@ public:
 			}
 
 			if(!found) {
-				return i;
+				return static_cast<size_t>(i);
 			}
 		}
